Move raiseHoneytokenAlert to a synced file so the AIM Cloud build resolves it
The honeytoken alert helper was defined in agent_service.go, which the AIM Cloud
sync treats as protected (the cloud keeps a divergent copy and skips the synced
version). capability_service.go, which the sync does copy, calls the helper, so
the cloud build failed with 'undefined: raiseHoneytokenAlert' on every deploy
since the honeytoken feature merged. Relocating the verbatim helper to a synced
file in the same package lets the cloud build resolve it. No logic change; the
public build is identical.
